Created ResponseHandler, renamed RequestHandler to Hyperdeck, Parser no longer asynchronous
There are TODO's in places.
Was only going to make a few tweaks then got carried away sorry :P
